- update minimum SFML version to 2.4 (fix deprecated get/setColor) - embed sfeMovie - update cmake to add FFmpeg FindFFMPEG.cmake - update deprecated code for FFMPEG3 upgrade in sfeMovie (90%) - remove unused code from sfeMovie (SubtitleStream)
- add isUsable to item
- add touch support (untested)
- add UIText interface
- execute action given drawable id and action name
- refactor buttons
- refactor cel decoding (fix small font)
- refactor input events code
- refactor setText actions into one (text.setText)
- refactor StringButton, Text2
- update android minimum version
- update itemClass description to support skipping matches
- update UIObject interface to add method to get action

- update gamefiles
- resize level panel if user panels are open/closed and if size = 640x480
- update item prefixes and descriptions
- update level right click

+++ src/Utils.h
@@ -62,7 +62,7 @@ namespace Utils
 		auto outputDiff = x > y ? x - y : y - x;
 
 		val -= inputRange.x;
-		val = std::lround((double)val * (double)outputDiff / (double)inputDiff) + (double)outputRange.x;
+		val = std::lround((double)val * (double)outputDiff / (double)inputDiff) + (long)outputRange.x;
 		return val;
 	}
